requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
lightgbm>=3.3.0
xgboost>=1.5.0

# Data processing
pyyaml>=6.0
requests>=2.25.0
beautifulsoup4>=4.9.0

# Visualization and apps
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
streamlit>=1.0.0

# API and web
fastapi>=0.70.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Statistical and optimization
statsmodels>=0.13.0
pymc>=4.0.0  # Optional for Bayesian extensions

# Development and testing
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0
jupyter>=1.0.0

# Optional for enhanced features
optuna>=2.10.0  # Hyperparameter optimization
shap>=0.40.0    # Model interpretability
